+++ modules/contracts/src.ts/tests/channelMastercopy.spec.ts
@@ -1,6 +1,6 @@
 import { expect } from "@connext/vector-utils";
 import { AddressZero, HashZero, Zero } from "@ethersproject/constants";
-import { Contract } from "ethers";
+import { Contract } from "@ethersproject/contracts";
 
 import { deployContracts } from "../actions";
 import { AddressBook } from "../addressBook";
@@ -27,6 +27,16 @@ describe("ChannelMastercopy", function() {
     const BalanceZero = [
       [Zero, Zero],
       [AddressZero, AddressZero],
+    ];
+    const WithdrawDataZero = [
+      AddressZero,
+      AddressZero,
+      AddressZero,
+      Zero,
+      Zero,
+      AddressZero,
+      Zero,
+      "0x"
     ];
     const CoreChannelStateZero = [AddressZero, AddressZero, AddressZero, [], [], [], [], Zero, Zero, HashZero, Zero];
     const CoreTransferStateZero = [
@@ -57,8 +67,8 @@ describe("ChannelMastercopy", function() {
       { name: "depositAlice", args: [AddressZero, Zero /*, HashZero */] },
 
       // from ICMCWithdraw
-      { name: "getWithdrawalTransactionRecord", args: [AddressZero, AddressZero, Zero, Zero] },
-      { name: "withdraw", args: [AddressZero, AddressZero, Zero, Zero, HashZero, HashZero] },
+      { name: "getWithdrawalTransactionRecord", args: [WithdrawDataZero] },
+      { name: "withdraw", args: [WithdrawDataZero, HashZero, HashZero] },
 
       // from ICMCAdjudicator
       { name: "getChannelDispute", args: [] },